+++ plydata/verbs.py
@@ -705,8 +705,7 @@ class summarize(DataOperator):
     - ``nth(x, n)`` - *nth* value of ``x`` or ``numpy.nan``.
     - ``n_distinct(x)`` - Number of distint elements in ``x``.
     - ``n_unique(x)`` - Alias of ``n_distinct``.
-    - ``{n}`` - Number of elements in current group. A special
-      function is created and substituted in place of ``{n}``.
+    - ``n()`` - Number of elements in current group.
 
     The aliases of the Numpy functions save you from typing 3 or 5 key
     strokes and you get better column names. i.e ``min(x)`` instead of
@@ -728,17 +727,17 @@ class summarize(DataOperator):
     2  2      4.0
     3  3      5.0
 
-    >>> df >> group_by('y') >> summarize(y_count='{n}')
+    >>> df >> group_by('y') >> summarize(y_count='n()')
        y  y_count
     0  0        2
     1  1        2
     2  2        1
     3  3        1
 
-    You can use ``{n}`` even when there are no groups.
+    You can use ``n()`` even when there are no groups.
 
-    >>> df >> summarize('{n}')
-       {n}
+    >>> df >> summarize('n()')
+       n()
     0    6
     """
     new_columns = None